requirements.txt
# ============================================================================
# Agno Backend API Service - 依赖包列表
# ============================================================================

# ============================================================================
# 核心框架依赖
# ============================================================================
# Agno 框架 - 多智能体系统框架（使用最新版本）
agno

# FastAPI - Web框架
fastapi>=0.104.0

# Uvicorn - ASGI服务器（包含标准扩展）
uvicorn[standard]>=0.24.0

# Pydantic - 数据验证和序列化
pydantic>=2.0.0

# ============================================================================
# HTTP客户端依赖
# ============================================================================
# httpx - 现代HTTP客户端（agno和API调用需要）
httpx>=0.25.0

# aiohttp - 异步HTTP客户端（可能被agno使用）
aiohttp>=3.9.0

# ============================================================================
# 模型API依赖
# ============================================================================
# openai - OpenAI SDK（用于OpenAIEmbedder和兼容API）
openai>=1.0.0

# ============================================================================
# 数据库相关依赖
# ============================================================================
# SQLAlchemy - ORM框架
sqlalchemy>=2.0.0

# PyMySQL - MySQL数据库驱动
pymysql>=1.1.0

# AsyncMy - 异步MySQL驱动
asyncmy>=0.2.0

# ============================================================================
# 向量数据库相关依赖
# ============================================================================
# PyMilvus - Milvus向量数据库客户端
pymilvus>=2.3.0

# ============================================================================
# 环境配置依赖
# ============================================================================
# python-dotenv - 环境变量管理
python-dotenv>=1.0.0

# PyYAML - YAML配置文件解析
pyyaml>=6.0.0

# ============================================================================
# 工具库依赖
# ============================================================================
# tenacity - 重试库（可能被agno使用）
tenacity>=8.2.0

# typing-extensions - 类型扩展（Python < 3.10需要）
typing-extensions>=4.8.0

# anyio - 异步IO库（FastAPI和uvicorn需要）
anyio>=4.0.0

# ============================================================================
# 测试依赖
# ============================================================================
# pytest - 单元测试框架
pytest>=7.0.0

# pytest-asyncio - 异步测试支持（FastAPI需要）
pytest-asyncio>=0.21.0

# ============================================================================
# 可选依赖（开发环境）
# ============================================================================
# 取消注释以启用开发工具
# black>=23.0.0              # 代码格式化工具
# isort>=5.12.0              # 导入排序工具
# flake8>=6.0.0              # 代码检查工具
# mypy>=1.0.0                # 类型检查工具

# ============================================================================
# OpenTelemetry 追踪依赖（可选，仅在启用 tracing 时需要）
# ============================================================================
# 如果需要使用 AgentOS 的 tracing 功能，取消注释以下依赖：
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
openinference-instrumentation-agno>=0.1.0

# ============================================================================
# Agno 扩展依赖（可选）
# ============================================================================
# 如果需要额外的MySQL支持功能，取消注释：
# agno[mysql]
